+++ BlockTrip_FE/src/App.jsx
@@ -1,13 +1,47 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-
 import routes from "./routes";
 import Home from "./pages/Home";
 import Travel from "./pages/Travel";
 import Blank from "./pages/Blank";
+import Loader from "./components/common/Loader";
+import { useState } from "react";
+import instance from "./services/index";
 
 function App() {
+  const [isLoading, setIsLoading] = useState(false);
+
+  instance.interceptors.request.use((config) => {
+    setIsLoading(true);
+    return config;
+  });
+
+  instance.interceptors.response.use(
+    function (response) {
+      setIsLoading(false);
+      return response;
+    },
+    function (error) {
+      if (error.response && error.response.status === 400) {
+        console.log("요청 변수 확인");
+      } else if (error.response && error.response.status === 401) {
+        console.log("인증 실패");
+      } else if (error.response && error.response.status === 403) {
+        console.log("호출 금지");
+      } else if (error.response && error.response.status === 404) {
+        console.log("API 없음");
+      } else if (error.response && error.response.status === 405) {
+        console.log("메서드 허용 안함");
+      } else if (error.response && error.response.status === 500) {
+        window.location.href = "/blank";
+      } else {
+        // 기타 에러 처리
+      }
+    }
+  );
+
   return (
     <BrowserRouter>
+      {isLoading && <Loader />}
       <Routes>
         <Route path={routes.home} element={<Home />} />
         <Route path={routes.result} element={<Travel />} />
